consolidate subscription services to single service
